server/requirements.txt
flask==3.1.2
flask-cors==6.0.2
gunicorn==21.2.0
pymongo==4.11.1
python-dotenv==1.0.1
